// src/shared/enums/ConsequenceType.ts
/**
 * @file ConsequenceType.ts
 * @author Mark Hutchison
 * An enum that details the different categories of consequence card types.
 */

/**
 * An enum detailing the exact options for consequence card types.
 */
export enum ConsequenceType {
  MoveForward = 0,
  MoveBackward = 1,
  LoseATurn = 2,
  SkipATurn = 3,
}